#1 Copyright. Controllers - videoCard, processors, memory.
