+++ src/main/java/com/multiverse/computer_parts/dto/videoCards/VideoCardDto.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright (c) 2022.
+ * This code was originally developed 03.07.2022, 20:31 by Valkov V.V.
+ */
+
 package com.multiverse.computer_parts.dto.videoCards;
 
 import java.util.List;
